EndDecode added
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@468 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/codec/WiredCodec.cpp
@@ -132,7 +132,7 @@ int WiredCodec::Encode(float **pcm, string OutExtension)
 	
 }
 
-int WiredCodec::Decode(const string &filename, t_Pcm *pcm, unsigned long length)
+unsigned long WiredCodec::Decode(const string &filename, t_Pcm *pcm, unsigned long length)
 {
 	list<t_WLib>::const_iterator		iterTWLib;
 	t_Pcm								mypcm;
